add schedule pages with semester
