--- frontend/src/components/ScheduleBlock.css
@@ -1,13 +1,0 @@
-.schedule-block {
-  display: flex;
-  padding: 10px;
-  justify-content: space-between;
-}
-
-.schedule-row {
-  display: flex;
-  gap: 10px;
-  width: 500px;
-  padding: 10px;
-  border: var(--border-size) var(--border-table) solid;
-}
