+++ frontend/src/components/SchedulerComponent.css
@@ -1,0 +1,32 @@
+.schedule-block {
+  display: flex;
+  padding: 10px;
+  justify-content: space-around;
+}
+
+.schedule-row {
+  box-sizing: border-box;
+  display: flex;
+  width: 482px;
+  gap: 5px;
+  padding: 10px;
+  border: var(--border-size) var(--border-table) solid;
+}
+
+.schedule-row div:nth-child(1) {
+  width: 80px;
+}
+
+.schedule-row div:nth-child(2) {
+  width: 350px;
+}
+
+.schedule-row div:nth-child(3) {
+  width: 20px;
+  text-align: center;
+}
+
+.schedule-draggable {
+  min-height: calc(5 * 41.5px);
+  width: 482px;
+}
